+++ ipscap/service/packet_filter.py
@@ -1,12 +1,19 @@
 import re
 
 from ipscap.util.raw_socket_entity import IPHeader
+from ipscap.configs import Constant
+import codecs
 
 
 class PacketFilter:
     def __init__(self, ev_parser):
-        self.trackings = []
-        self.ev_parser = ev_parser
+        self._find_bytes = None
+        self._trackings = []
+        self._ev_parser = ev_parser
+
+    def initialize(self, args):
+        if args.find:
+            self.prepare_find(args)
 
     def verify_capture(self, ip_header, protocol_header, args):
         is_capture = self.filter_packet(ip_header, protocol_header, args)
@@ -17,7 +24,7 @@ class PacketFilter:
             if is_capture:
                 tracking = (ip_header.src_ip, ip_header.dest_ip, port)
 
-                if tracking not in self.trackings:
+                if tracking not in self._trackings:
                     self._add_transfer(ip1, ip2, port)
             else:
                 if self._is_tracking_transfer(ip1, ip2, port):
@@ -36,12 +43,12 @@ class PacketFilter:
     def _add_transfer(self, ip1, ip2, port):
         tracking = (ip1, ip2, port)
 
-        self.trackings.append(tracking)
+        self._trackings.append(tracking)
 
     def _is_tracking_transfer(self, ip1, ip2, port):
         is_tracking = False
 
-        for tracking in self.trackings:
+        for tracking in self._trackings:
             if tracking[0] == ip1 and tracking[1] == ip2 and tracking[2] == port:
                 is_tracking = True
                 break
@@ -61,7 +68,7 @@ class PacketFilter:
         if not self.verify_find(ip_header, protocol_header, args):
             return False
 
-        if not self.ev_parser.is_empty():
+        if not self._ev_parser.is_empty():
             if not self.verify_condition(ip_header, protocol_header):
                 return False
 
@@ -99,85 +106,100 @@ class PacketFilter:
 
         return True
 
+    def prepare_find(self, args):
+        find_mode = args.find_mode
+
+        if find_mode == Constant.FIND_REGEX or find_mode == Constant.FIND_MATCH:
+            self._find_bytes = args.find.encode('utf-8')
+        elif find_mode == Constant.FIND_BINARY:
+            self._find_bytes = self.create_bytes_by_binary(args.find)
+        elif find_mode == Constant.FIND_HEX:
+            self._find_bytes = self.create_bytes_by_hex(args.find)
+
     def verify_find(self, ip_header, protocol_header, args):
+        find_mode = args.find_mode
+
         if args.find:
-            if not args.find_case_sensitive:
-                flags = re.IGNORECASE
-            else:
-                flags = 0
+            if find_mode == Constant.FIND_REGEX or find_mode == Constant.FIND_MATCH:
+                flags = re.IGNORECASE if find_mode == Constant.FIND_REGEX else 0
 
-            if not re.search(args.find.encode('utf-8'), protocol_header.payload_data, flags):
-                return False
+                if not re.search(self._find_bytes, protocol_header.payload_data, flags):
+                    return False
+            elif find_mode == Constant.FIND_BINARY:
+                transfer_data = ip_header.header_data + protocol_header.header_data + protocol_header.payload_data
 
-        if args.find_hex:
-            search_byte = self.create_byte_by_hex(args.find_hex)
+                if self._find_bytes not in transfer_data:
+                    return False
+            elif find_mode == Constant.FIND_HEX:
+                transfer_data = ip_header.header_data + protocol_header.header_data + protocol_header.payload_data
 
-            transfer_data = ip_header.header_data + protocol_header.header_data + protocol_header.payload_data
-
-            if search_byte not in transfer_data:
-                return False
+                if self._find_bytes not in transfer_data:
+                    return False
 
         return True
 
     def verify_condition(self, ip_header, protocol_header):
-        if self.ev_parser.assigned('port'):
+        if self._ev_parser.assigned('port'):
             port = protocol_header.dest_port if protocol_header.dest_port < protocol_header.src_port else protocol_header.src_port
 
-            if not self.ev_parser.evaluate('port', port):
+            if not self._ev_parser.evaluate('port', port):
                 return False
 
-        if self.ev_parser.assigned('client_port'):
+        if self._ev_parser.assigned('client_port'):
             port = protocol_header.dest_port if protocol_header.dest_port > protocol_header.src_port else protocol_header.src_port
 
-            if not self.ev_parser.evaluate('client_port', port):
+            if not self._ev_parser.evaluate('client_port', port):
                 return False
 
-        if self.ev_parser.assigned('src_port'):
-            if not self.ev_parser.evaluate('src_port', protocol_header.src_port):
+        if self._ev_parser.assigned('src_port'):
+            if not self._ev_parser.evaluate('src_port', protocol_header.src_port):
                 return False
 
-        if self.ev_parser.assigned('dest_port'):
-            if not self.ev_parser.evaluate('dest_port', protocol_header.dest_port):
+        if self._ev_parser.assigned('dest_port'):
+            if not self._ev_parser.evaluate('dest_port', protocol_header.dest_port):
                 return False
 
-        if self.ev_parser.assigned('ttl'):
-            if not self.ev_parser.evaluate('ttl', ip_header.ttl):
+        if self._ev_parser.assigned('ttl'):
+            if not self._ev_parser.evaluate('ttl', ip_header.ttl):
                 return False
 
         if ip_header.protocol == IPHeader.PROTOCOL_TCP:
-            if self.ev_parser.assigned('flags'):
-                flags = self.ev_parser.get_value('flags')
+            if self._ev_parser.assigned('flags'):
+                flags = self._ev_parser.get_value('flags')
 
                 if not (protocol_header.flags & flags):
                     return False
 
-            if self.ev_parser.assigned('seq'):
-                if not self.ev_parser.evaluate('seq', protocol_header.seq_no):
+            if self._ev_parser.assigned('seq'):
+                if not self._ev_parser.evaluate('seq', protocol_header.seq_no):
                     return False
 
-            if self.ev_parser.assigned('ack'):
-                if not self.ev_parser.evaluate('ack', protocol_header.ack_no):
+            if self._ev_parser.assigned('ack'):
+                if not self._ev_parser.evaluate('ack', protocol_header.ack_no):
                     return False
 
-            if self.ev_parser.assigned('window'):
-                if not self.ev_parser.evaluate('window', protocol_header.window):
+            if self._ev_parser.assigned('window'):
+                if not self._ev_parser.evaluate('window', protocol_header.window):
                     return False
 
-            if self.ev_parser.assigned('mss'):
-                if not self.ev_parser.evaluate('mss', protocol_header.tcp_options.get('mss')):
+            if self._ev_parser.assigned('mss'):
+                if not self._ev_parser.evaluate('mss', protocol_header.tcp_options.get('mss')):
                     return False
 
-            if self.ev_parser.assigned('wscale'):
-                if not self.ev_parser.evaluate('wscale', protocol_header.tcp_options.get('wscale')):
+            if self._ev_parser.assigned('wscale'):
+                if not self._ev_parser.evaluate('wscale', protocol_header.tcp_options.get('wscale')):
                     return False
 
-            if self.ev_parser.assigned('sack'):
-                if not self.ev_parser.evaluate('sack', protocol_header.tcp_options.get('sack')):
+            if self._ev_parser.assigned('sack'):
+                if not self._ev_parser.evaluate('sack', protocol_header.tcp_options.get('sack')):
                     return False
 
         return True
 
-    def create_byte_by_hex(self, data):
+    def create_bytes_by_binary(self, data):
+        return codecs.decode(data, 'unicode_escape').encode('latin1')
+
+    def create_bytes_by_hex(self, data):
         data = re.sub(r'\s', '', data)
 
         try: