+++ ipsurv/core/target_parser.py
@@ -15,7 +15,7 @@ import ipaddress
 class TargetParser(ABC):
     """
     Description:
-    https://deer-hunt.github.io/ipsurv/pages/program_architecture_classes.html#targetparser
+    https://deer-hunt.github.io/ipsurv/pages/ipsurv-cmd/program_architecture_classes.html#targetparser
     """
     def __init__(self, args, pipeline, dns_resolver):
         """