+++ ipsurv/requester/http.py
@@ -15,7 +15,7 @@ class HttpRequester(Requester):
     :type timeout: float
 
     Description:
-    https://deer-hunt.github.io/ipsurv/pages/program_architecture_classes.html#requester
+    https://deer-hunt.github.io/ipsurv/pages/ipsurv-cmd/program_architecture_classes.html#requester
     """
 
     def __init__(self, timeout=None):