+++ ipsurv/requester/ip_info.py
@@ -13,7 +13,7 @@ class IpInfoRequester(Requester):
     :type token: str
 
     Description:
-    https://deer-hunt.github.io/ipsurv/pages/program_architecture_classes.html#requester
+    https://deer-hunt.github.io/ipsurv/pages/ipsurv-cmd/program_architecture_classes.html#requester
     """
 
     def __init__(self, timeout=None, token=None):