+++ ipsurv/requester/server_reactivity.py
@@ -12,7 +12,7 @@ class ServerReactivity(Requester):
     :type timeout: float
 
     Description:
-    https://deer-hunt.github.io/ipsurv/pages/program_architecture_classes.html#requester
+    https://deer-hunt.github.io/ipsurv/pages/ipsurv-cmd/program_architecture_classes.html#requester
     """
 
     def __init__(self, timeout=4.0):