+++ ipsurv/serializer/json_serializer.py
@@ -6,7 +6,7 @@ from ipsurv.serializer.serializer import Serializer
 class JsonSerializer(Serializer):
     """
     Description:
-    https://deer-hunt.github.io/ipsurv/pages/program_architecture_classes.html#serializer
+    https://deer-hunt.github.io/ipsurv/pages/ipsurv-cmd/program_architecture_classes.html#serializer
     """
 
     def __init__(self, args):