+++ ipsurv/serializer/serializer.py
@@ -8,7 +8,7 @@ from ipsurv.core.entity import ValueData
 class Serializer(ABC):
     """
     Description:
-    https://deer-hunt.github.io/ipsurv/pages/program_architecture_classes.html#serializer
+    https://deer-hunt.github.io/ipsurv/pages/ipsurv-cmd/program_architecture_classes.html#serializer
     """
 
     def __init__(self, args):